Start of binding stuff, EndpointReference utils, Transport changes
Updated schemas to add the generateElementClass setting to generate 
the element classes so that xsd:any will actually unmarshal as objects

Added EndpointReferenceUtils for extracting things like WSDL's and
Ports from EndpointReferenceType objects.   This will need to be
updated as the ws-addressing people figure out the override 
order and such.

Updated client transport API's to handle some level of async messages.

Start of Binding API's for client side and a base Binding class
that can be used for a starting point.




git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@434632 13f79535-47bb-0310-9956-ffa450edef68