Add release packaging
Create celtix-src.zip to contain a release source distribution.  Basically, a copy of our repository.
Create celtix-bin.zip to contain a binary release distribution.  This needs to be filled in more once we figure out what goes in it and add things like env scripts, command line bat/sh scripts, etc....
Update some utilities in the ant scripts to work on unix.
Update WSDLManagerImpl to stop printing stuff in the the unit tests.
Update AutomaticWorkQueueTest to less likely fail on an overloaded machine.




git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@434690 13f79535-47bb-0310-9956-ffa450edef68