Fix SOAP binding server side parsing issues, headers...
Change AbstractServerBinding to dispatch the message directly on the transport supplied thread.   The transport was closing the streams before the other thread was reading it.  

Update all the Context creations to "wrap" any existing contexts instead of creating all new ones so values in the exiting context are propgated.

Updated SOAP message parsing to grab the HTTP header when available so things like mime boundaries and such are available.



git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@434700 13f79535-47bb-0310-9956-ffa450edef68