Move checkstyle to after compile
Move checkstyle to after the javac step so that various other tests that require the .class files can be enabled.
Enable the IllegalThrows and RedundantThrows checks in checkstyle.  (there are too many issues with IllegalCatch right now)
Collapse the multiple source javac commands down to a single javac (now that jaxb fixed the XMLRootElement stuff)


git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@434742 13f79535-47bb-0310-9956-ffa450edef68