Added in some more extended types to the type tests.
There's a couple jaxb ri issues for which a few of the tests have been
commented out:
rpc-literal tests of the xsd:date types cause a ClassCastException in
the jaxb code.
For a couple of the enumeration types - ColourEnum for instance - the
generated fromValue() method isn't declared static.
Restrictions in the schema aren't enforced, e.g., maxLength.

And couple of other things to note:
The ri generated code flattens nested struct/choice in the schema
to the toplevel class.  However the spec seems to say that the default
should be to generate the nested classes.
There are certain types in which the generated method signatures for the
doc-literal and rpc-literal styles differ (use List vs array).



git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@434965 13f79535-47bb-0310-9956-ffa450edef68