Update perf test...
Update server to not print the entire message to the screen.  That's bad for performance.
Update the base to warm up the jit for 5 secs before testing.
Update common_build to turn off assertions (bad for performance) and give it more memory so less garbage collection (bad for performance)

Drops average response time on my machine from 16.5ms to 10.5ms.  (must be careful of exactly what you are performance testing...)


git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@435010 13f79535-47bb-0310-9956-ffa450edef68