add ResourceManager to load resources.
ResourceManager delegates to ResourceResolvers which can be configured on the Bus or added programmatically.  The ResourceManager can be accessed as a service of the bus or as a singleton that is used before the Bus or configuration sub-system are initialized.
Update Bus to initialize a ResourceManager.
Replaced calls to getSystemResourceAsStream with calls to ResourceManager.getResourceAsAsStream
Update ResourceInjector to use ResourceManager to resolve resources.

git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@435029 13f79535-47bb-0310-9956-ffa450edef68