Merging John Lifter's soap_header sample "as-is" with just enough reformatting to avoid PMD style errors breaking the kit build.
No time for it now, but this demo could do with a bit of re-work,
to make its origins as a system test less obvious.

In particular, reduce the number of header types to something less
intimidating (like 1 or 2), change the *test* identifiers in the WSDL, 
and add some made-up purpose to the demo (e.g. headers containing a 
session cookie or auth token). 



git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@435064 13f79535-47bb-0310-9956-ffa450edef68