Fixed <a href="http://forge.objectweb.org/tracker/index.php?func=detail&aid=304952&group_id=192&atid=350241"> bug304952 </a>
The problem is caused by the lack of checking the duplicated importing schemas.
But the testcase attached with the snag still fail because of the JAXB bugs.
To track the issue, please see the <a href="http://forge.objectweb.org/tracker/index.php?func=detail&aid=304953&group_id=192&atid=350241"> bug304953 </a>

git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@435732 13f79535-47bb-0310-9956-ffa450edef68