workaround for Axis Interop test failure on OS X
It's not really an OSX failure -- more of a slow Java/machine failure.  There appears to be 
some timing related issue between Axis and Jetty whereby an Axis client is having problems 
when a two-way invocation is directly followed by a one-way.  The Axis client intermittently 
fails to correctly read the response from the server served up by Jetty. 



git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@435909 13f79535-47bb-0310-9956-ffa450edef68