Allow resource handlers to initialised via @Resource injection
The JAXWS spec is not clear about handler initialisation.   The init method is not in the Handler interface anymore (was in previous revision) and there is a fuddled mention of injection.  This support is therefore disabled by default for User defined handlers.

This is the behaviour when enabled.
If there is an init method, it will be invoked.  If there is no init method, the init-param values specified in the handler-config will be injected into appropriately annotated members of the handler class.  When all values have been injected, any methods decorated with @PostConstruct will be called.

- turned off handler initialisation by injection or init method by default.  Can be enabled in config.
- add InitParamResourceResolver for resolving init params from handler config 
- add init of handlers using @resource injection.  Will happen if enabled in config and no init method is present on handler.
- alllow ResourceManager to resolve resource from specified list of resolvers.
- trim handler class name and init params to remove extra white space

git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@435933 13f79535-47bb-0310-9956-ffa450edef68