- Start of a JaxWsServiceFactoryBean for the server side of the JAX-WS hello world. The bean uses the WSDLServiceFactory in conjunction with the jax ws annotations to build up a Service. Endpoint.publish should be able to use this to help publish a Service. Additionally its nicely reusable for people not wanting to use the nasty Endpoint.publish() method. Still needs work yet though...
- removed the JAX-WS dependency from the SOAP binding.
We were throwing a WebServiceException. I changed this to a
RuntimeException for now, but we really need a FaultException
which contains things like the fault code.

- Changed checkstyle back to allow ifs without brackets. I really
don't feel like this rule is representative of code quality and I 
don't feel its worth the time to go and change my code.


git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@436366 13f79535-47bb-0310-9956-ffa450edef68