CXF-117  Fix for wsdl validation with imported wsdls/xsds.
The validators shouldn't choke on valid wsdl which happens to use schema
types imported from another wsdl or xsd file.
Added test for validating a wsdl which includes a wsdl.
Added test for validating a wsdl which includes an xsd.
Fixed some typos in error messages.
Anyone familiar with the validators please feel free to review and comment.


git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@450565 13f79535-47bb-0310-9956-ffa450edef68