CXF-258
Fixed NEP in the RPC style when we don't have the WSDLLocation set in the WebService annotation
Also added the testcase for it.

Previous SEI in tests are generated in the testutils, so it always contains the WSDLLocation in the annotation.
If there is no WSDLLocation it's a different branch which we didn't cover by the test.
The issue was caused by wrong processing of wrapped style.


git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@478071 13f79535-47bb-0310-9956-ffa450edef68