[CXF-290] Work around for URIResolver issue with spaces in pathname.
Spaces in the pathname on windows caused problems in the URIResolver (space
characters need to be escaped in a URI).  Added a test and made the same
sort of change for a couple of similar issues in the validator tool.  There's
still a couple of related things in the wsdl2java tool which I'll track down
later.
Also fixed a couple of typos.


git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@482822 13f79535-47bb-0310-9956-ffa450edef68