Minor fix to HttpUriMapper.
If the path part of the uri passed in to the getResourceBase() method had only a
single '/', then it would return "/" instead of the full resource base.  Caused
a problem in the case where a service publishes two endpoints on addresses like
this:
http://localhost:8080/SoapPort1
http://localhost:8080/SoapPort2
The first published endpoint would get all the messages.  Updated the
HttpUriMapperTest.


git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@495793 13f79535-47bb-0310-9956-ffa450edef68