Start of migration to support Spring 2.0 NamespaceHandlers instead of our own property editor mechanism. This allows a much more compact syntax:
    <h:conduit id="{urn:test:ns}Foo.http-conduit">
      <h:client ConnectionTimeout="97"/>
    </h:conduit>

It also means that people won't have to use our custom ApplicationContext.
As part of this I consolidated the conduit and configuration schemas into the 
http-conf schema. Otherwise people would have to use multiple namespaces to 
write their configuration which was annoying and confusing.

I'm not sure I'm happy with our configuration syntax though. It has mix 
conventions for naming things (some start with an upper case letter and some
do not). Also, it still seems overly verbose/redundat. Not sure what the
best strategy is here though.

Also, I changed the SpringBusFactory/BusApplicationContext so we didn't refresh()
twice as the logging was a bit annoying.


git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@501982 13f79535-47bb-0310-9956-ffa450edef68