Create an "Object binding". This allows local invocation of services without any serialization. The binding and operation are selected via properties on the message. I.e.
message.put(ObjectBinding.BINDING, new QName("someBinding"));
message.put(ObjectBinding.OPERATION, new QName("someOperation"));

We could maybe enhance this to work with URIs like James was suggesting
as well.

I made a few slight changes to the local transport to support this.
There is a DIRECT_DISPATCH flag which just tells the local transport
not to use streams, but instead to copy the content from the sent message
to a new message which is passed on to a Destination.




git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@524475 13f79535-47bb-0310-9956-ffa450edef68