Flesh out the Feature concept in CXF and try to bring together Features & Policy.
I added 3 Feature implementations in this commit:
- LoggingFeature. This simply adds the logging interceptors to the specifed endpoint.
- WSAddressingFeature
- WSPolicyFeature

The most interesting of these is probably the WSPolicyFeature. This allows inlining
of WS-Policy inside Spring configurations. This eliminates the need for an external 
policy file when using Policy. For instance:

<endpoint>
  <features>
    <Policy><UsingAddressing/></Policy>
  </features>
<endpoint>

Or you can also reference policies in a Spring config and create <beans>:

<Policy wsu:Id="FooPolicy">...</Policy>

<endpoint>
  <features>
    <ref bean="FooPolicy"/>
    <bean class="org.apache.cxf.feature.LoggingFeature"/>
  </features>
<endpoint>

One of the things this made painfully clear (after an hour or two of banging 
my head against the desk), is that we need to have different spring IDs for
both <endpoint> and <client>. Otherwise they clash. So I changed the syntax to 
this:

<jaxws:client id="...FooPort" createdFromAPI="true"/>
<jaxws:server id="...FooPort" createdFromAPI="true"/>

What this does is signal to the DefinitionParser that we should append
".jaxws-client" or ".jaxws-endpoint" to the id internally. I would love a
better name than  "createdFromAPI" if you have any ideas. I didn't 
automatically append ".jaxws-foo" when abstract=true because that would
break "pure spring" users who are doing parent/child relationships with
their beans.



git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@531126 13f79535-47bb-0310-9956-ffa450edef68