CXF-781: Fix a problem with Aegis whereby it didn't detect that both an ArrayType and a ListType represented the same schema type, and the schema was duplicated inside the WSDL. (Amazingly, this is a bug with XFire too - I can't believe that no one has run into it before! I guess people just don't mix arrays and Lists together very often?) Many thanks to Benson Margulies for the excellent unit test that he provided
CXF-784: Set the content-type on the outgoing fault chain inside the HTTP Binding.

Add in a StaxDataBinding implementation which makes it possible to write services
which read straight from the XMLStreamReader. You can also return an XMLStreamReader
or an XMLStreamWriterCallback to write the response. This is still kind of hackish,
I'd like to see if I could make this work without removing the DocLiteral/Bare/Wrapper
Interceptors. But I need to start a discussion on how to do that on the mailing list.




git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@556305 13f79535-47bb-0310-9956-ffa450edef68