Fix up support for Aegis exceptions. This mostly involved adding a FaultOutInterceptor and making Aegis/W3CDOMStreamReader work with the latest DepthXMLStreamReader change. I also added a hook to pass in the actual Fault object to Aegis so it could build the Exception appropriately.
CXF-811: Fix the problem with missing types in the WSDL for Aegis.

Fix a problem where argument names were arg00 with the simple frontend.

You may notice I added a FaultInfoException class as well. This makes it possible to
use databindings like JAXB which don't support serializing exceptions, with the simple
frontend. This work is ongoing....
 


git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@557765 13f79535-47bb-0310-9956-ffa450edef68