CXF-785: Fix an NPE if a namespace prefix is invalid.
CXF-800: I fixed the root cause of the bug in the JRA project, but I 
added an additional unit test here. Basically the issue was we weren't
able to handle resources like /customer/{123} and /customer/details/{123}.
The first would always get triggered.

We're dependent on a JRA snapshot until I finallize a release tomorrow.



git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@559631 13f79535-47bb-0310-9956-ffa450edef68