CXF-846: Remove the charset=utf-8 from start-info in MTOM Content-Type because tomcat barfs on it.
CXF-749: Wire in the HTTPSession for the servlet/jetty transports. Will add a unit test for this shortly.

Improve fault code parsing since it was creating fault codes like null:Foo.

Change the impl generation in wsdl2java to not call class.getPackage() as that doesn't work on all classloaders.




git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@559856 13f79535-47bb-0310-9956-ffa450edef68